fix(warnings): comparison between signed and unsigned
- cast signed to unsigned if always >= 0, since unsigned is wider
- or just replace the type with unsigned if that's cleaner

+++ ImageLounge/src/DkGui/DkWidgets.cpp
@@ -238,7 +238,7 @@ void DkThumbsSaver::thumbLoaded()
     mNumSaved++;
     emit numFilesSignal(mNumSaved);
 
-    if (mNumSaved == mWatchers.size() || mStop) {
+    if ((unsigned int)mNumSaved == mWatchers.size() || mStop) {
         if (mPd) {
             mPd->close();
             mPd->deleteLater();
